Conflicts error fixed
