+++ app.rb
@@ -1,15 +1,17 @@
-require './classes/item'
-require './classes/book'
-require './classes/label'
-require './classes/game'
-require './classes/source'
-require './classes/author'
+# app.rb
+require_relative './classes/item'
+require_relative './classes/book'
+require_relative './classes/label'
+require_relative './classes/game'
+require_relative './classes/source'
+require_relative './classes/author'
 
 class App
   def initialize
     @books = []
     @labels = []
     @authors = []
+    @games = []
   end
 
   # list methods
@@ -19,7 +21,7 @@ class App
       puts 'No books found.'
     else
       @books.each do |book|
-        puts "ID: #{book.id}, Author: #{book.author.first_name}, Publisher: #{book.publisher},
+        puts "ID: #{book.id}, Author: #{book.author.full_name}, Publisher: #{book.publisher},
         Cover state: #{book.cover_state}"
       end
     end
@@ -36,6 +38,28 @@ class App
     end
   end
 
+  def list_all_games
+    puts "\nListing all games:"
+    if @games.empty?
+      puts 'No games found.'
+    else
+      @games.each do |game|
+        puts "#{game.id}. #{game.label} (Genre: #{game.genre}, Author: #{game.author.full_name})"
+      end
+    end
+  end
+
+  def list_all_authors
+    puts "\nListing all authors:"
+    if @authors.empty?
+      puts 'No authors found.'
+    else
+      @authors.each do |author|
+        puts "ID: #{author.id}, Full Name: #{author.full_name}, Items Count: #{author.items.count}"
+      end
+    end
+  end
+  
   # add methods
   def add_book
     puts 'Published by:'
@@ -63,6 +87,32 @@ class App
     choose_author(book)
 
     @books << book
+  end
+
+  def add_game
+    puts '===== Add a Game ====='
+    print 'Enter genre: '
+    genre = gets.chomp
+    print 'Enter author (full name): '
+    author_name = gets.chomp
+    author = Author.find_by_full_name(author_name) || Author.new(first_name: author_name.split.first, last_name: author_name.split.last)
+    print 'Enter source: '
+    source = gets.chomp
+    print 'Enter label: '
+    label = gets.chomp
+    print 'Enter publish date (YYYY-MM-DD): '
+    publish_date = gets.chomp
+    print 'Is it multiplayer? (true/false): '
+    multiplayer = gets.chomp.downcase == 'true'
+    print 'Enter last played date (YYYY-MM-DD): '
+    last_played_at = gets.chomp
+
+    game = Game.new(genre, author, source, label, publish_date, multiplayer, last_played_at)
+    game.move_to_archive if game.can_be_archived?
+
+    @games << game
+
+    puts 'Game added successfully!'
   end
 
   private
